+++ packages/ckeditor5-dev-tests/lib/tasks/runautomatedtests.js
@@ -104,6 +104,23 @@ function createEntryFile( globPatterns, production ) {
 	const entryFileContent = allFiles
 		.map( file => 'import "' + file + '";' );
 
+	// Inject the custom chai assertions. See ckeditor/ckeditor5#9668.
+	const assertionsDir = path.join( __dirname, '..', 'utils', 'automated-tests', 'assertions' ).replace( /\\/g, '/' );
+	const customAssertions = fs.readdirSync( assertionsDir ).map( assertionFileName => {
+		return [
+			assertionFileName,
+			path.parse( assertionFileName ).name.replace( /-([a-z])/g, value => value[ 1 ].toUpperCase() )
+		];
+	} );
+
+	// Two loops are needed to achieve correct order in `ckeditor5/build/.automated-tests/entry-point.js`.
+	for ( const [ fileName, functionName ] of customAssertions ) {
+		entryFileContent.push( `import ${ functionName }Factory from "${ assertionsDir }/${ fileName }";` );
+	}
+	for ( const [ , functionName ] of customAssertions ) {
+		entryFileContent.push( `${ functionName }Factory( chai );` );
+	}
+
 	if ( production ) {
 		entryFileContent.unshift( `
 const originalWarn = console.warn;