+++ packages/ckeditor5-dev-utils/lib/translations/cleanpofilecontent.js
@@ -12,9 +12,11 @@ const PO = require( 'pofile' );
  * containing information where to add new translations or fix the existing ones.
  *
  * @param {String} poFileContent Content of the translation file.
+ * @param {Object} [options={}]
+ * @param {Boolean} [options.simplifyLicenseHeader] Whether to skip adding the contribute URL in the header.
  * @returns {String}
  */
-module.exports = function cleanPoFileContent( poFileContent ) {
+module.exports = function cleanPoFileContent( poFileContent, options = {} ) {
 	const po = PO.parse( poFileContent );
 
 	// Remove personal data from headers.
@@ -24,21 +26,30 @@ module.exports = function cleanPoFileContent( poFileContent ) {
 		'Plural-Forms': po.headers[ 'Plural-Forms' ]
 	};
 
+	const copyright = po.comments.find( comment => comment.includes( 'Copyright' ) );
+
 	// Clean comments.
 	po.comments = [
-		po.comments.find( comment => comment.includes( 'Copyright' ) ),
-		'',
 		'                                    !!! IMPORTANT !!!',
 		'',
 		'        Before you edit this file, please keep in mind that contributing to the project',
 		'               translations is possible ONLY via the Transifex online service.',
-		'',
-		'        To submit your translations, visit https://www.transifex.com/ckeditor/ckeditor5.',
-		'',
-		'                  To learn more, check out the official contributor\'s guide:',
-		'    https://ckeditor.com/docs/ckeditor5/latest/framework/guides/contributing/contributing.html',
 		''
 	];
+
+	if ( copyright ) {
+		po.comments.unshift( copyright, '' );
+	}
+
+	if ( !options.simplifyLicenseHeader ) {
+		po.comments.push(
+			'        To submit your translations, visit https://www.transifex.com/ckeditor/ckeditor5.',
+			'',
+			'                  To learn more, check out the official contributor\'s guide:',
+			'    https://ckeditor.com/docs/ckeditor5/latest/framework/guides/contributing/contributing.html',
+			''
+		);
+	}
 
 	return po.toString();
 };